# backend/src/main/java/com/example/sessionauth/dto/SubdivisionDTO.java
package com.example.sessionauth.dto;

import lombok.AccessLevel;
import lombok.Data;
import lombok.experimental.FieldDefaults;

@Data
@FieldDefaults(level = AccessLevel.PRIVATE)
public class SubdivisionDTO {

    Long id;

    String name;
}