# backend/src/main/resources/swagger.yaml
openapi: 3.0.4
info:
  title: Session Auth API - Document Management
  description: API for managing documents in the Session Authentication system
  version: 1.0.0
  contact:
    email: support@example.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api.example.com/api/v1
    description: Production server
...
paths:
  /subdivisions/create:
    post:
      tags:
        - Subdivision
      summary: Create a new subdivision
      description: Creates a new subdivision with the provided details. Requires ADMIN authority.
      operationId: createSubdivision
      security:
        - bearerAuth: [ADMIN]
      requestBody:
        description: Subdivision to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubdivisionDTO'
      responses:
        '201':
          description: Subdivision created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubdivisionDTO'
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /subdivisions/{id}:
    get:
      tags:
        - Subdivision
      summary: Get subdivision by ID
      description: Returns a single subdivision by its ID. Requires ADMIN authority.
      operationId: getSubdivisionById
      security:
        - bearerAuth: [ADMIN]
      parameters:
        - name: id
          in: path
          description: ID of subdivision to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubdivisionDTO'
        '400':
          description: Invalid ID supplied
        '404':
          description: Subdivision not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Subdivision
      summary: Update an existing subdivision
      description: Updates an existing subdivision. Requires ADMIN authority.
      operationId: updateSubdivision
      security:
        - bearerAuth: [ADMIN]
      parameters:
        - name: id
          in: path
          description: ID of subdivision to update
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Subdivision data to update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubdivisionDTO'
      responses:
        '200':
          description: Subdivision updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubdivisionDTO'
        '400':
          description: Invalid ID or input
        '404':
          description: Subdivision not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Subdivision
      summary: Delete a subdivision
      description: Deletes a subdivision by ID. Requires ADMIN authority.
      operationId: deleteSubdivision
      security:
        - bearerAuth: [ADMIN]
      parameters:
        - name: id
          in: path
          description: ID of subdivision to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Subdivision deleted successfully
        '400':
          description: Invalid ID supplied
        '404':
          description: Subdivision not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /subdivisions/getAll:
    get:
      tags:
        - Subdivision
      summary: Get all subdivisions
      description: Returns a list of all subdivisions. Requires ADMIN authority.
      operationId: getAllSubdivisions
      security:
        - bearerAuth: [ADMIN]
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SubdivisionDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /students/create:
    post:
      tags:
        - Student
      summary: Create a new student
      description: Creates a new student with the provided details. Requires EMPLOYEE authority.
      operationId: createStudent
      security:
        - bearerAuth: [EMPLOYEE]
      requestBody:
        description: Student to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentDTO'
      responses:
        '201':
          description: Student created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentDTO'
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /students/{id}:
    get:
      tags:
        - Student
      summary: Get student by ID
      description: Returns a single student by its ID. Requires EMPLOYEE authority.
      operationId: getStudentById
      security:
        - bearerAuth: [EMPLOYEE]
      parameters:
        - name: id
          in: path
          description: ID of student to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentDTO'
        '400':
          description: Invalid ID supplied
        '404':
          description: Student not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Student
      summary: Update an existing student
      description: Updates an existing student. Requires EMPLOYEE authority.
      operationId: updateStudent
      security:
        - bearerAuth: [EMPLOYEE]
      parameters:
        - name: id
          in: path
          description: ID of student to update
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Student data to update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentDTO'
      responses:
        '200':
          description: Student updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentDTO'
        '400':
          description: Invalid ID or input
        '404':
          description: Student not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Student
      summary: Delete a student
      description: Deletes a student by ID. Requires EMPLOYEE authority.
      operationId: deleteStudent
      security:
        - bearerAuth: [EMPLOYEE]
      parameters:
        - name: id
          in: path
          description: ID of student to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Student deleted successfully
        '400':
          description: Invalid ID supplied
        '404':
          description: Student not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /students/getAll:
    get:
      tags:
        - Student
      summary: Get all students
      description: Returns a list of all students. Requires EMPLOYEE authority.
      operationId: getAllStudents
      security:
        - bearerAuth: [EMPLOYEE]
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StudentDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    SubdivisionDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
          description: Subdivision ID
        name:
          type: string
          example: "Human Resources"
          description: Name of the subdivision
      required:
        - name
...
    StudentDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
          description: Student ID
        person:
          $ref: '#/components/schemas/PersonDTO'
        yearStarted:
          type: integer
          example: 2021
          description: Year the student started their studies
        financialForm:
          type: string
          enum:
            - TARGETING
            - BUDGET
            - PAYMENT
          example: "BUDGET"
          description: The financial form for the student
      required:
        - person
        - yearStarted
        - financialForm
...
    PersonDTO:
      type: object
      properties:
        name:
          type: string
          example: "John"
          description: First name
        surname:
          type: string
          example: "Doe"
          description: Last name
        patronymic:
          type: string
          example: "Smith"
          description: Middle name (optional)
          nullable: true
        phone:
          type: string
          example: "+1234567890"
          description: Phone number
...
    Error:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          example: "2023-10-01T12:00:00Z"
        status:
          type: integer
          example: 400
        error:
          type: string
          example: "Bad Request"
        message:
          type: string
          example: "Invalid input data"
        path:
          type: string
          example: "/api/v1/documents"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication